requirements.txt
Flask
python-dotenv
Flask-Talisman
